finish rename module 'declarative' -> 'standard_tests'
