--- airbyte_cdk/test/declarative/__init__.py
@@ -1,6 +1,0 @@
-# Copyright (c) 2024 Airbyte, Inc., all rights reserved.
-"""Declarative tests framework.
-
-This module provides fixtures and utilities for testing Airbyte sources and destinations
-in a declarative way.
-"""
