+++ airbyte_cdk/test/standard_tests/declarative_sources.py
@@ -9,9 +9,9 @@ from boltons.typeutils import classproperty
 from airbyte_cdk.sources.declarative.concurrent_declarative_source import (
     ConcurrentDeclarativeSource,
 )
-from airbyte_cdk.test.declarative.models import ConnectorTestScenario
-from airbyte_cdk.test.declarative.utils.job_runner import IConnector
+from airbyte_cdk.test.standard_tests._job_runner import IConnector
 from airbyte_cdk.test.standard_tests.connector_base import MANIFEST_YAML
+from airbyte_cdk.test.standard_tests.models import ConnectorTestScenario
 from airbyte_cdk.test.standard_tests.source_base import SourceTestSuiteBase
