+++ airbyte_cdk/test/standard_tests/models/__init__.py
@@ -1,4 +1,4 @@
-from airbyte_cdk.test.declarative.models.scenario import (
+from airbyte_cdk.test.standard_tests.models.scenario import (
     ConnectorTestScenario,
 )
